move StoreLocatorProvider to app level to ensure selectedStore is saved in the context
