+++ packages/template-retail-react-app/app/pages/store-locator/index.jsx
@@ -12,33 +12,24 @@ import {Box, Container} from '@salesforce/retail-react-app/app/components/shared
 import Seo from '@salesforce/retail-react-app/app/components/seo'
 import StoreLocatorContent from '@salesforce/retail-react-app/app/components/store-locator-modal/store-locator-content'
 
-// Others
-import {
-    StoreLocatorContext,
-    useStoreLocator
-} from '@salesforce/retail-react-app/app/components/store-locator-modal/index'
-
 const StoreLocator = () => {
-    const storeLocator = useStoreLocator()
 
     return (
-        <StoreLocatorContext.Provider value={storeLocator}>
-            <Box data-testid="store-locator-page" bg="gray.50" py={[8, 16]}>
-                <Seo title="Store Locator" description="Find a Store" />
-                <Container
-                    overflowY="scroll"
-                    paddingTop={8}
-                    width={['90%']}
-                    bg="white"
-                    paddingBottom={14}
-                    marginTop={8}
-                    marginBottom={8}
-                    borderRadius="base"
-                >
-                    <StoreLocatorContent />
-                </Container>
-            </Box>
-        </StoreLocatorContext.Provider>
+        <Box data-testid="store-locator-page" bg="gray.50" py={[8, 16]}>
+            <Seo title="Store Locator" description="Find a Store" />
+            <Container
+                overflowY="scroll"
+                paddingTop={8}
+                width={['90%']}
+                bg="white"
+                paddingBottom={14}
+                marginTop={8}
+                marginBottom={8}
+                borderRadius="base"
+            >
+                <StoreLocatorContent />
+            </Container>
+        </Box>
     )
 }
